extract workflow state routes
